fix: create service interface
